refactor(ui-react): extract rive workflow preview
